copypasting from match ups log; "guardians were having problems in closed maps, not being able to defend against a massive orc spam. not killing in time"
Yep, old change, but wasn't implemented because it is actually a buff (190% from 165% against forwrooks and archery range). It allows orcs to take down faster my beatiful forgeworks, that's why i had removed it. Now i re-added it because i want clear scalars on orc weapons. 180% global, 190% dwarves, 150% elves isen.

It is sure that eventually the 150% scalar changes to 155%