new discovery. difference between ANY and NONE
ANY +HEROE +STRUCTURE = means it works on structures and also heroes.
NONE +HEROE +STRUCTURE = means it will work on a object with BOTH this kind ofs. Won't work on heroes, won't work on structures.

This is tested and means a lot of 1.06 scalars never worked, like this.
LegolasHawkstrike's "10% NONE +STRUCTURE +MordorCatapult +IsengardBallista +GondorTrebuchet +GondorTrebuchetWall" =  broken.
Lots of 1.09 scalars, and the worst part of everything, a lot of scalars/filters of beta since i fixed the crash at Exit button.
easy to fix since we code thinking it works with the "ANY" mode, but fixing it globaly means "enable" some broken 1.6/1.9 scalars. the hard work is to log it.