-citadel -undermine
i have been trying to make citadel castable above units. I found out how but it means the expansions pad might not spawn if during the cast something is on the zone. i will see if i can make the shockwave clear the area, then delay the spawn of the expansion sets.

-fixed the retardeness of one-hitting any machine in 500000 radius.. now just radius cursor, where you know what will die and what not.