-Dain audio fix
Ok this is huge change or let's be honest; "fix2.

the audio settigns changes in 109v1 didn't rlly fix 560 zoom not playing half of the sounds, especial the battles ones, unless you zoom in. It only removed the fading out which doesn't apply for most of the bfme1ish sounds becasue they are coded for 300 zoom.
so i revised every sound and applied a formula based on 450 zoom for minimun range sounds, and 560 zoom for max range sounds.

for example: a vanilla sound like "min100max350"
-i consider it like "Min" being -200 than vanilla zoom and "Max" being +50 than vanilla zoom)
Then i make the "Min" to be -200 than Beta450zoom, and the "Max" to be +50 than Beta560zoom)
Equals to Min:250 Max:610 from Min:100 Max:350.

I applied it on sounds with values inferior to 560

and i re-edited audiosettings.ini since some fading out settings are missing since 09.

What do you think?